rename: (DashboardFragment & DashboardViewModel) -> (MapFragment & MapViewModel)
