--- app/src/main/java/com/moroom/android/ui/navui/dashboard/DashboardViewModel.kt
@@ -1,7 +1,0 @@
-package com.moroom.android.ui.navui.dashboard
-
-import androidx.lifecycle.ViewModel
-
-class DashboardViewModel : ViewModel() {
-
-}
