+++ app/src/main/java/com/moroom/android/ui/navui/map/MapViewModel.kt
@@ -1,0 +1,7 @@
+package com.moroom.android.ui.navui.map
+
+import androidx.lifecycle.ViewModel
+
+class MapViewModel : ViewModel() {
+
+}
