+++ app/src/main/res/layout/fragment_map.xml
