+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,9 +7,9 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
+        android:id="@+id/navigation_map"
         android:icon="@drawable/location_on_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_map" />
 
     <item
         android:id="@+id/navigation_profile"
