+++ app/src/main/res/navigation/nav_graph.xml
@@ -12,23 +12,23 @@
         tools:layout="@layout/fragment_home" >
 
         <action
-            android:id="@+id/action_navigation_home_to_navigation_dashboard"
-            app:destination="@id/navigation_dashboard" />
+            android:id="@+id/action_navigation_home_to_navigation_map"
+            app:destination="@id/navigation_map" />
         <action
             android:id="@+id/action_navigation_home_to_navigation_profile"
             app:destination="@id/navigation_profile" />
     </fragment>
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.moroom.android.ui.navui.dashboard.DashboardFragment"
-        android:label="@string/dash_fragment_label"
-        tools:layout="@layout/fragment_dashboard" >
+        android:id="@+id/navigation_map"
+        android:name="com.moroom.android.ui.navui.map.MapFragment"
+        android:label="@string/map_fragment_label"
+        tools:layout="@layout/fragment_map" >
         <action
-            android:id="@+id/action_navigation_dashboard_to_navigation_profile2"
+            android:id="@+id/action_navigation_map_to_navigation_profile"
             app:destination="@id/navigation_profile" />
         <action
-            android:id="@+id/action_navigation_dashboard_to_navigation_home"
+            android:id="@+id/action_navigation_map_to_navigation_home"
             app:destination="@id/navigation_home" />
     </fragment>
 
@@ -42,7 +42,7 @@
             android:id="@+id/action_navigation_profile_to_navigation_home"
             app:destination="@id/navigation_home" />
         <action
-            android:id="@+id/action_navigation_profile_to_navigation_dashboard"
-            app:destination="@id/navigation_dashboard" />
+            android:id="@+id/action_navigation_profile_to_navigation_map"
+            app:destination="@id/navigation_map" />
     </fragment>
 </navigation>
